requirements.txt
django
djangorestframework
djangorestframework-simplejwt
drf-yasg
python-dotenv
gunicorn